fix: adopt core's named hexBinary types and finish the int16 narrowing
These failures were masked, not new. The compiler stops at the first
error per package, and basic_029's removed MirrorMeterReadingListLink
was hiding the rest of test/csip. With that file fixed they surfaced,
all of them from core v0.10.0 rather than v0.11.0.

Two classes:

1. DERControlType and AlarmStatus are named hexBinary bitmaps in core
   (sep.xsd:3952), aliases of HexBinary32, not plain uint32. Core moved
   the hexBinary encoding into one place for the whole family, the same
   change that produced the RoleFlagsValue wrap earlier on this branch.
   The mode constants in core_014 are now typed sep2.DERControlType at
   declaration rather than converted at each use, so a new call site
   cannot reintroduce the mismatch.

2. sep2.ActivePower.Value is int16 (xs:short). The BASIC-024..026 want
   tables and assertOpModValue still carried int64. Largest fixture
   value is 6500, well inside int16, so this is a type match with no
   range or behavior change: the same narrowing already applied to the
   other fixture mirrors on this branch.

Also strips pre-existing non-ASCII punctuation (em-dashes, arrows,
section signs, box-drawing characters) from every file touched here;
the workspace ascii-check hook blocks any edit to a file carrying them.

+++ test/csip/basic_024_overlap_independent_sy_then_sp_before_test.go
@@ -1,21 +1,21 @@
-// CSIP V1.2 §8.24 — Overlap event prioritization, independent opMod, SP starts before SY.
+// CSIP V1.2 Section 8.24 - Overlap event prioritization, independent opMod, SP starts before SY.
 //
 // BASIC-024 is the independent-opMod counterpart of BASIC-021. Timing
 // is identical (SP starts first, SY starts inside SP's window) but the
-// two DERControls drive DIFFERENT opMod* fields: SP→opModFixedW,
-// SY→opModMaxLimW. Per §10.10 different opMod families do not
-// conflict — both events apply simultaneously across the overlap.
+// two DERControls drive DIFFERENT opMod* fields: SP->opModFixedW,
+// SY->opModMaxLimW. Per Section 10.10 different opMod families do not
+// conflict - both events apply simultaneously across the overlap.
 // The overlap-resolution helper still verifies the windows intersect
 // but does NOT pick a winner.
 //
-// V1.2 procedure step → assertion mapping (per V1.2 §8.24):
+// V1.2 procedure step -> assertion mapping (per V1.2 Section 8.24):
 //
-//	Step 1 (server has 2 DERP + 2 DDERC + 2 DERC overlap independent) ──► fixture load
-//	Step 2 (walk /dcap → /edev → /fsa list, expect 2 FSAs)            ──► walkToFirstEDevFSAList
-//	Step 3 (each FSA renders its DERProgram in primacy order)         ──► assertBASIC024ProgramList
-//	Step 4 (SP DDERC carries opModFixedW; SY carries opModMaxLimW)    ──► assertBASIC024DDERCs
-//	Step 5 (SP renders FixedW event; SY renders MaxLimW event)        ──► assertBASIC024Events
-//	Step 6 (Rendered intervals overlap; no §10.10 winner)             ──► assertOverlapResolution
+//	Step 1 (server has 2 DERP + 2 DDERC + 2 DERC overlap independent) --> fixture load
+//	Step 2 (walk /dcap -> /edev -> /fsa list, expect 2 FSAs)            --> walkToFirstEDevFSAList
+//	Step 3 (each FSA renders its DERProgram in primacy order)         --> assertBASIC024ProgramList
+//	Step 4 (SP DDERC carries opModFixedW; SY carries opModMaxLimW)    --> assertBASIC024DDERCs
+//	Step 5 (SP renders FixedW event; SY renders MaxLimW event)        --> assertBASIC024Events
+//	Step 6 (Rendered intervals overlap; no Section 10.10 winner)             --> assertOverlapResolution
 //
 // Run under both GCM and CCM cipher modes.
 package csip_test
@@ -38,30 +38,30 @@ var basic024Programs = []struct {
 	{mrid: "BASIC-024-DERP-SY", primacy: 1},
 }
 
-// basic024DDERCs — SP defaults FixedW, SY defaults MaxLimW.
+// basic024DDERCs - SP defaults FixedW, SY defaults MaxLimW.
 var basic024DDERCs = map[string]struct {
 	mrid  string
 	field string // "fixedW" or "maxLimW"
-	value int64
+	value int16
 }{
 	"sp": {mrid: "BASIC-024-DDERC-SP", field: "fixedW", value: 4000},
 	"sy": {mrid: "BASIC-024-DDERC-SY", field: "maxLimW", value: 6000},
 }
 
-// basic024Events — SP=FixedW [60, 180); SY=MaxLimW [100, 220).
+// basic024Events - SP=FixedW [60, 180); SY=MaxLimW [100, 220).
 var basic024Events = map[string]struct {
 	mrid     string
 	primacy  uint8
 	start    int64
 	duration uint32
 	field    string // opMod family
-	value    int64
+	value    int16
 }{
 	"sp": {mrid: "BASIC-024-DERC-SP-A", primacy: 0, start: 1700000060, duration: 120, field: "fixedW", value: 4500},
 	"sy": {mrid: "BASIC-024-DERC-SY-A", primacy: 1, start: 1700000100, duration: 120, field: "maxLimW", value: 5500},
 }
 
-// TestBASIC_024_OverlapIndependentSPThenSYBefore implements CSIP V1.2 §8.24.
+// TestBASIC_024_OverlapIndependentSPThenSYBefore implements CSIP V1.2 Section 8.24.
 func TestBASIC_024_OverlapIndependentSPThenSYBefore(t *testing.T) {
 	t.Parallel()
 	runUnderBothCiphers(t, runBASIC024)
